+++ src/macrowhisper/Watcher/RecordingsFolderWatcher.swift
@@ -30,7 +30,7 @@ class RecordingsFolderWatcher {
         self.historyManager = historyManager
         self.socketCommunication = socketCommunication
         self.triggerEvaluator = TriggerEvaluator(logger: logger)
-        self.clipboardMonitor = ClipboardMonitor(logger: logger)
+        self.clipboardMonitor = ClipboardMonitor(logger: logger, preRecordingBufferSeconds: configManager.config.defaults.clipboardBuffer)
         self.actionExecutor = ActionExecutor(logger: logger, socketCommunication: socketCommunication, configManager: configManager, clipboardMonitor: clipboardMonitor)
         self.versionChecker = versionChecker